implemented better search and eval

+++ libparallax/core/bitboard.hpp
@@ -34,8 +34,8 @@ constexpr auto square_to_bit(const square target_square) noexcept -> bitboard {
   return bitboard{1} << static_cast<std::int32_t>(target_square);
 }
 
-constexpr auto popcount(const bitboard board) noexcept -> std::uint32_t {
-  return static_cast<std::uint32_t>(std::popcount(board));
+constexpr auto popcount(const bitboard board) noexcept -> std::int32_t {
+  return std::popcount(board);
 }
 
 constexpr auto lsb(const bitboard board) noexcept -> square {
